+++ packages/flame_tiled/lib/src/tiled_component.dart
@@ -15,6 +15,7 @@ class TiledComponent extends Component {
 
   @override
   void render(Canvas canvas) {
+    super.render(canvas);
     tileMap.render(canvas);
   }