+++ packages/flame/CHANGELOG.md
@@ -4,6 +4,7 @@
  - `changePriority` no longer breaks game loop iteration
  - Move component mixin checks to their own files
  - Fix exception when game was rebuilt
+ - Add `@mustCallSuper` on `Component.render`
 
 ## [1.0.0-releasecandidate.15]
  - Fix issue with `Draggable`s not being removed from `draggables` list